+++ src/validate/mod.rs
@@ -1,5 +1,7 @@
 pub mod slip_verify;
 pub mod true_money_slip_verify;
+pub mod bcel_one_proof;
 
 pub use slip_verify::{slip_verify, SlipVerifyData};
 pub use true_money_slip_verify::{true_money_slip_verify, TrueMoneySlipVerifyData};
+pub use bcel_one_proof::{bcel_one_proof, BcelOneProofData};