+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-12-15
+
+### Added
+- Implement auto fix bad CRC for slip verify validator
+- Added BCEL OneProof validator
+
 ## [1.0.1] - 2025-09-08
 
 ### Documentation